# api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "anvilml-server",
    "description": "",
    "license": {
      "name": ""
    },
    "version": "0.1.27"
  },
  "paths": {
...
    "/v1/artifacts": {
      "get": {
        "tags": [
          "Artifacts"
        ],
        "summary": "List artifacts",
        "description": "Lists artifact metadata, optionally filtered by job ID.",
        "operationId": "list_artifacts",
        "parameters": [
          {
            "name": "job_id",
            "in": "query",
            "description": "Optional job UUID to filter artifacts — only artifacts from this job are returned.\n`None` returns all artifacts regardless of job.",
            "required": false,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of artifacts",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ArtifactMeta"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Database error"
          }
        }
      }
    },
    "/v1/artifacts/{hash}": {
      "get": {
        "tags": [
          "Artifacts"
        ],
        "summary": "Get artifact",
        "description": "Serves raw PNG bytes for a content-addressed artifact by its SHA-256 hash.",
        "operationId": "get_artifact",
        "parameters": [
          {
            "name": "hash",
            "in": "path",
            "description": "SHA-256 content hash of the artifact",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Raw PNG bytes",
            "content": {
              "image/png": {}
            }
          },
          "404": {
            "description": "Artifact not found"
          },
          "500": {
            "description": "I/O error"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ArtifactMeta": {
        "type": "object",
        "description": "Metadata for a generated, content-addressed PNG artifact.\n\nThis struct captures the identity, generation parameters, and\nstorage location of a single output artifact produced by a worker.\nThe `hash` field is the SHA-256 hex content address used by\n`anvilml-artifacts` as its primary key.",
        "required": [
          "hash",
          "job_id",
          "width",
          "height",
          "seed",
          "steps",
          "created_at",
          "file_path"
        ],
        "properties": {
          "created_at": {
            "type": "string",
            "format": "date-time",
            "description": "Timestamp when the artifact was created."
          },
          "file_path": {
            "type": "string",
            "description": "Filesystem path to the saved PNG file."
          },
          "hash": {
            "type": "string",
            "description": "SHA-256 hex content address — primary key for artifact storage."
          },
          "height": {
            "type": "integer",
            "format": "int32",
            "description": "Generated image height in pixels.",
            "minimum": 0
          },
          "job_id": {
            "type": "string",
            "format": "uuid",
            "description": "UUID of the job that produced this artifact."
          },
          "seed": {
            "type": "integer",
            "format": "int64",
            "description": "Random seed used for generation (i64 to support negative seeds)."
          },
          "steps": {
            "type": "integer",
            "format": "int32",
            "description": "Number of diffusion steps used in generation.",
            "minimum": 0
          },
          "width": {
            "type": "integer",
            "format": "int32",
            "description": "Generated image width in pixels.",
            "minimum": 0
          }
        }
      },
...
    }
  },
...
}